+++ src/comp/Pipeline.cpp
@@ -32,7 +32,7 @@ bool Pipeline::create(const base::Manifest& vertexManifest, const base::Shape* s
     std::vector<VkVertexInputAttributeDescription> vertexAttributes(vertexManifest.numberOfElements());
     for (size_t i = 0; i < vertexManifest.numberOfElements(); ++i) {
         vertexAttributes[i].binding = 0;
-        vertexAttributes[i].location = i;
+        vertexAttributes[i].location = static_cast<uint32_t>(i);
         VkFormat format;
         switch (vertexManifest.typeForElement(i)) {
         case base::Manifest::ElementType::kFloat:
@@ -63,7 +63,7 @@ bool Pipeline::create(const base::Manifest& vertexManifest, const base::Shape* s
     vertexInputInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_STATE_CREATE_INFO;
     vertexInputInfo.vertexBindingDescriptionCount = 1;
     vertexInputInfo.pVertexBindingDescriptions = &vertexBindingDescription;
-    vertexInputInfo.vertexAttributeDescriptionCount = vertexAttributes.size();
+    vertexInputInfo.vertexAttributeDescriptionCount = static_cast<uint32_t>(vertexAttributes.size());
     vertexInputInfo.pVertexAttributeDescriptions = vertexAttributes.data();
 
     // Input Assembly
@@ -173,7 +173,7 @@ bool Pipeline::create(const base::Manifest& vertexManifest, const base::Shape* s
     if (pushConstantBlockSize) {
         pushConstantRange.stageFlags = VK_SHADER_STAGE_FRAGMENT_BIT | VK_SHADER_STAGE_VERTEX_BIT;
         pushConstantRange.offset = 0;
-        pushConstantRange.size = pushConstantBlockSize;
+        pushConstantRange.size = static_cast<uint32_t>(pushConstantBlockSize);
         pipelineLayoutInfo.pushConstantRangeCount = 1;
         pipelineLayoutInfo.pPushConstantRanges = &pushConstantRange;
     } else {
@@ -254,7 +254,7 @@ bool Pipeline::createCompute(std::shared_ptr<vk::Shader> computeShader, VkDescri
     if (pushConstantBlockSize) {
         pushConstantRange.stageFlags = VK_SHADER_STAGE_COMPUTE_BIT;
         pushConstantRange.offset = 0;
-        pushConstantRange.size = pushConstantBlockSize;
+        pushConstantRange.size = static_cast<uint32_t>(pushConstantBlockSize);
         pipelineLayoutInfo.pushConstantRangeCount = 1;
         pipelineLayoutInfo.pPushConstantRanges = &pushConstantRange;
     } else {
